requirements.txt
python-telegram-bot==20.5
dropbox==12.1.0
requests==2.31.0